Alot ffs

+++ keyboard/init.sh
@@ -1,1 +1,3 @@
-cp ~/.config/keyboard/ABCsw.keylayout ~/Library/Keyboard\ Layouts/
+if [ ! -e ~/Library/Keyboard\ Layouts/ABCsw.keylayout ]; then
+  cp ~/.config/keyboard/ABCsw.keylayout ~/Library/Keyboard\ Layouts/
+fi
